# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Wallet API created by Kiraly Mate
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
securityDefinitions:
  ApiKeyAuth:
    type: apiKey
    in: header
    name: session_id
security:
  - ApiKeyAuth: []
...
paths:
  /register:
    x-swagger-router-controller: register
    post:
      description: Signs up a new user and saves the credentials + the account number
      operationId: signUpUser
      security: []
      tags:
      - "Register"
      parameters:
        - name: user
          in: body
          description: The new user
          required: true
          schema:
            $ref: '#/definitions/NewUser'
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/NewUserResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  NewUser:
    required:
      - username
      - pin
      - accountNumber
    properties:
      username:
        type: string
        example: 'user'
      pin:
        type: integer
        example: 1234
      accountNumber:
        type: string
        example: '11111111-22222222-33333333'
  NewUserResponse:
    required:
      - sessionId
    properties:
      sessionId:
        type: string
        example: 'abc12345'
...